fix longstanding bug preventing rc4 decryption #9
Obviously RC4 is not a block cipher, so it will have to be handled
slightly different. There are more than just a few issues (both in the
Cuckoo repository as well as the httpreplay repository) reported
regarding this bug - now it has finally been fixed.
Thanks to dfirbuddy for a PCAP file identifying this issue.