replace custom netlib http parser by our internal methods
It might be that this commit breaks some functionality. If that's the
case, we should certainly revert [1] and verify the changes in
behavior / output. On the plus side, this leaves us with just one
implementation of the HTTP/HTTPS parsing which is useful since we just
fixed a bug where httpreplay would attempt to parse non-decrypted
HTTPS/TLS traffic as HTTP due to missing TLS Master Secrets (this would
mostly just end up logging lots of incorrect and useless warnings).

[1]: git checkout 36dbef6dfcfc02a6eb4a2749c32e15aeee97f322